Move several actions to ParallelAlgorithms/Actions
FunctionsOfTimeAreReady
LimiterActions
FilterAction
