+++ src/Evolution/Executables/Burgers/CMakeLists.txt
@@ -12,6 +12,7 @@ add_spectre_executable(
 target_link_libraries(
   ${EXECUTABLE}
   PRIVATE
+  Actions
   Burgers
   BurgersAnalyticData
   BurgersSolutions
