+++ src/Evolution/Executables/ForceFree/CMakeLists.txt
@@ -12,6 +12,7 @@ add_spectre_executable(
 target_link_libraries(
   ${EXECUTABLE}
   PRIVATE
+  Actions
   Charmxx::main
   DiscontinuousGalerkin
   DomainCreators
