+++ src/Evolution/Executables/RadiationTransport/M1Grey/CMakeLists.txt
@@ -12,6 +12,7 @@ add_spectre_executable(
 target_link_libraries(
   ${EXECUTABLE}
   PRIVATE
+  Actions
   Charmxx::main
   CoordinateMaps
   DiscontinuousGalerkin
