+++ src/Evolution/Executables/RelativisticEuler/Valencia/CMakeLists.txt
@@ -2,6 +2,7 @@
 # See LICENSE.txt for details.
 
 set(LIBS_TO_LINK
+  Actions
   Charmxx::main
   CoordinateMaps
   DiscontinuousGalerkin
