+++ src/Evolution/Executables/ScalarAdvection/CMakeLists.txt
@@ -2,6 +2,7 @@
 # See LICENSE.txt for details.
 
 set(LIBS_TO_LINK
+  Actions
   Charmxx::main
   DgSubcell
   DiscontinuousGalerkin
