+++ src/Evolution/Executables/ScalarTensor/CMakeLists.txt
@@ -12,6 +12,7 @@ add_spectre_executable(
 target_link_libraries(
   ${EXECUTABLE}
   PRIVATE
+  Actions
   ApparentHorizonFinder
   Charmxx::main
   ControlSystem
