+++ tests/Unit/Evolution/DiscontinuousGalerkin/Limiters/CMakeLists.txt
@@ -25,6 +25,7 @@ add_test_library(${LIBRARY} "${LIBRARY_SOURCES}")
 target_link_libraries(
   ${LIBRARY}
   PRIVATE
+  Actions
   Boost::boost
   Limiters
   )
