+++ tests/Unit/NumericalAlgorithms/LinearOperators/CMakeLists.txt
@@ -21,6 +21,7 @@ add_test_library(${LIBRARY} "${LIBRARY_SOURCES}")
 target_link_libraries(
   ${LIBRARY}
   PRIVATE
+  Actions
   DataStructures
   DiscontinuousGalerkin
   Domain
